# web_ui/openapi.json
{"openapi":"3.1.0","info":{"title":"susutaku backend","description":"","license":{"name":"MIT","url":"https://opensource.org/licenses/MIT"},"version":"0.1.0"},"paths":{"/api/chat":{"post":{"tags":[],"operationId":"chat","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatReply"}}}},"500":{"description":"","content":{"text/plain":{"schema":{"type":"string"}}}}}}},"/api/health":{"get":{"tags":[],"operationId":"health","responses":{"200":{"description":"","content":{"text/plain":{"schema":{"type":"string"}}}}}}},"/api/models":{"get":{"tags":[],"operationId":"models","responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ModelInfo"}}}}}}}},"/api/models/select":{"post":{"tags":[],"operationId":"select_model","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SelectRequest"}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SelectReply"}}}},"400":{"description":"","content":{"text/plain":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"ChatReply":{"type":"object","required":["reply","searched","tokenizer","prompt_tokens","prompt_tps","decode_tokens","decode_tps"],"properties":{"decode_tokens":{"type":"integer","minimum":0},"decode_tps":{"type":"number","format":"double"},"model":{"type":["string","null"]},"prompt_tokens":{"type":"integer","minimum":0},"prompt_tps":{"type":"number","format":"double"},"reply":{"type":"string"},"searched":{"type":"boolean"},"tokenizer":{"type":"string"}}},"ChatRequest":{"type":"object","required":["message"],"properties":{"max_tokens":{"type":["integer","null"],"minimum":0},"message":{"type":"string"},"search":{"type":["string","null"],"description":"\"on\" | \"off\" | \"auto\" (default: model decides, Zed-style)."},"think":{"type":["boolean","null"],"description":"Emit a reasoning block (default: off — closed thinking)."},"tokenizer":{"type":["string","null"],"description":"\"normal\" | \"katgpt\" (default: normal)."}}},"ModelInfo":{"type":"object","required":["name","loadable","bytes","engine","selected"],"properties":{"bytes":{"type":"integer","format":"int64","minimum":0},"engine":{"type":"string"},"loadable":{"type":"boolean"},"name":{"type":"string"},"selected":{"type":"boolean"}}},"SelectReply":{"type":"object","required":["selected"],"properties":{"selected":{"type":"string"}}},"SelectRequest":{"type":"object","required":["name"],"properties":{"name":{"type":"string"}}}}}}