# web_ui/src/api/generated/model/chatReply.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * susutaku backend
 * OpenAPI spec version: 0.1.0
 */
import type { ChatReplyModel } from './chatReplyModel';

export interface ChatReply {
  /** @minimum 0 */
  decode_tokens: number;
  decode_tps: number;
  model?: ChatReplyModel;
  /** @minimum 0 */
  prompt_tokens: number;
  prompt_tps: number;
  reply: string;
  searched: boolean;
  tokenizer: string;
}