# web_ui/src/api/generated/model/chatRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * susutaku backend
 * OpenAPI spec version: 0.1.0
 */
import type { ChatRequestMaxTokens } from './chatRequestMaxTokens';
import type { ChatRequestSearch } from './chatRequestSearch';
import type { ChatRequestThink } from './chatRequestThink';
import type { ChatRequestTokenizer } from './chatRequestTokenizer';

export interface ChatRequest {
  /** @minimum 0 */
  max_tokens?: ChatRequestMaxTokens;
  message: string;
  /** "on" | "off" | "auto" (default: model decides, Zed-style). */
  search?: ChatRequestSearch;
  /** Emit a reasoning block (default: off — closed thinking). */
  think?: ChatRequestThink;
  /** "normal" | "katgpt" (default: normal). */
  tokenizer?: ChatRequestTokenizer;
}