# web_ui/src/api/generated/model/modelInfo.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * susutaku backend
 * OpenAPI spec version: 0.1.0
 */

export interface ModelInfo {
  /** @minimum 0 */
  bytes: number;
  engine: string;
  loadable: boolean;
  name: string;
  selected: boolean;
}